[調整]Titleの描画クラス内で使用していない変数の削除・メモリの解放

+++ UniverSports_Remake/DirectX12 MainProject/ResultScene.cpp
@@ -15,7 +15,7 @@
 // Initialize member variables.
 ResultScene::ResultScene()
 {
-	DontDestroy->NowScene_	= (int)NextScene::ResultScene;
+	DontDestroy->NowScene_	= NextScene::ResultScene;
 
 	descriptorHeap_			= nullptr;
 	spriteBatch_			= nullptr;
